fix more clang warnings

+++ src/actions.c
@@ -3789,13 +3789,13 @@ action_runtime_delete_main_playlist_items(void) {
     ncm_error_clear(&error);
     for (int32 i = 0; i < count; i += 1) {
         if (!ncm_mpd_client_delete(&global_mpd, positions[i], &error)) {
-            free2(positions, (uint64)count * SIZEOF(*positions));
+            free2(positions, count*SIZEOF(*positions));
             ncm_song_array_destroy(&songs);
             return action_runtime_mpd_error(&error);
         }
     }
 
-    free2(positions, (uint64)count * SIZEOF(*positions));
+    free2(positions, count*SIZEOF(*positions));
     ncm_song_array_destroy(&songs);
     (void)ncm_status_update_full(&global_mpd, NULL, &error);
     ncm_statusbar_print_cstring((int32)Config.message_delay_time,
@@ -3845,14 +3845,14 @@ action_runtime_delete_playlist_editor_items(void) {
     for (int32 i = 0; i < count; i += 1) {
         if (!ncm_mpd_client_playlist_delete(&global_mpd, playlist.path,
                                             positions[i], &error)) {
-            free2(positions, (uint64)count * SIZEOF(*positions));
+            free2(positions, count * SIZEOF(*positions));
             ncm_playlist_destroy(&playlist);
             ncm_song_array_destroy(&songs);
             return action_runtime_mpd_error(&error);
         }
     }
 
-    free2(positions, (uint64)count * SIZEOF(*positions));
+    free2(positions, count * SIZEOF(*positions));
     ncm_playlist_destroy(&playlist);
     ncm_song_array_destroy(&songs);
     native_playlist_editor_screen_request_content_update(screen);
@@ -4190,7 +4190,7 @@ action_runtime_move_main_playlist_items(NcmSongArray *songs, bool down) {
         global_mpd.command_list_active = false;
     }
 
-    free2(positions, (uint64)count * SIZEOF(*positions));
+    free2(positions, count * SIZEOF(*positions));
     if (!success) {
         return action_runtime_mpd_error(&error);
     }
@@ -4216,7 +4216,7 @@ action_runtime_move_stored_playlist_items(NcmSongArray *songs, bool down) {
     ncm_playlist_init(&playlist);
     success = native_playlist_editor_screen_current_playlist(screen, &playlist);
     if (!success) {
-        free2(positions, (uint64)count * SIZEOF(*positions));
+        free2(positions, count * SIZEOF(*positions));
         ncm_playlist_destroy(&playlist);
         return false;
     }
@@ -4247,7 +4247,7 @@ action_runtime_move_stored_playlist_items(NcmSongArray *songs, bool down) {
         global_mpd.command_list_active = false;
     }
 
-    free2(positions, (uint64)count * SIZEOF(*positions));
+    free2(positions, count * SIZEOF(*positions));
     ncm_playlist_destroy(&playlist);
     if (!success) {
         return action_runtime_mpd_error(&error);
@@ -4345,19 +4345,19 @@ action_runtime_move_main_playlist_items_to(void) {
         }
         song = nc_menu_item_at(menu, NC_MENU_ITEMS_ALL, i);
         if (song == NULL) {
-            free2(positions, (uint64)item_count * SIZEOF(*positions));
+            free2(positions, item_count * SIZEOF(*positions));
             return false;
         }
         positions[count++] = ncm_song_position(song);
     }
     if (count <= 0) {
-        free2(positions, (uint64)item_count * SIZEOF(*positions));
+        free2(positions, item_count * SIZEOF(*positions));
         return false;
     }
 
     action_runtime_sort_positions(positions, count, false);
     if ((target >= positions[0]) && (target <= positions[count - 1])) {
-        free2(positions, (uint64)item_count * SIZEOF(*positions));
+        free2(positions, item_count * SIZEOF(*positions));
         return true;
     }
 
@@ -4382,7 +4382,7 @@ action_runtime_move_main_playlist_items_to(void) {
     if (!success && global_mpd.command_list_active) {
         global_mpd.command_list_active = false;
     }
-    free2(positions, (uint64)item_count * SIZEOF(*positions));
+    free2(positions, item_count * SIZEOF(*positions));
     if (!success) {
         return action_runtime_mpd_error(&error);
     }
@@ -4445,13 +4445,13 @@ action_runtime_move_playlist_editor_items_to(void) {
         }
         song = nc_menu_item_at(menu, NC_MENU_ITEMS_ALL, i);
         if (song == NULL) {
-            free2(positions, (uint64)item_count * SIZEOF(*positions));
+            free2(positions, item_count * SIZEOF(*positions));
             return false;
         }
         positions[count++] = ncm_song_position(song);
     }
     if (count <= 0) {
-        free2(positions, (uint64)item_count * SIZEOF(*positions));
+        free2(positions, item_count * SIZEOF(*positions));
         return false;
     }
 
@@ -4459,14 +4459,14 @@ action_runtime_move_playlist_editor_items_to(void) {
     success = native_playlist_editor_screen_current_playlist(screen, &playlist);
     if (!success) {
         ncm_playlist_destroy(&playlist);
-        free2(positions, (uint64)item_count * SIZEOF(*positions));
+        free2(positions, item_count * SIZEOF(*positions));
         return false;
     }
 
     action_runtime_sort_positions(positions, count, false);
     if ((target >= positions[0]) && (target <= positions[count - 1])) {
         ncm_playlist_destroy(&playlist);
-        free2(positions, (uint64)item_count * SIZEOF(*positions));
+        free2(positions, item_count * SIZEOF(*positions));
         return true;
     }
 
@@ -4495,7 +4495,7 @@ action_runtime_move_playlist_editor_items_to(void) {
     }
 
     ncm_playlist_destroy(&playlist);
-    free2(positions, (uint64)item_count * SIZEOF(*positions));
+    free2(positions, item_count * SIZEOF(*positions));
     if (!success) {
         return action_runtime_mpd_error(&error);
     }
